requirements/requirements.txt
decorator==4.0.6
Flask==0.10.1
Flask-Login==0.3.2
Flask-Migrate==2.0.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Uploads==0.2.0
Flask-WTF==0.12
gunicorn==19.6.0
infinity==1.4
inflect==0.2.5
intervals==0.8.0
itsdangerous==0.24
Jinja2==2.8
markdown2==2.3.0
MarkupSafe==0.23
nodeenv==0.13.6
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
python-dateutil==2.5.3
simplegeneric==0.8.1
six==1.10.0
sqlacodegen==1.1.6
SQLAlchemy==1.0.11
SQLAlchemy-Utils==0.32.9
traitlets==4.0.0
validators==0.10.3
Werkzeug==0.11.3
WTForm==1.0
WTForms==2.1
WTForms-Components==0.10.0
